% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_airport_sounding.R
\name{download_airport_sounding}
\alias{download_airport_sounding}
\title{Download sounding from Jorge Chavez Airport
Download sounding from Jorge Chavez Airport (SPIM or 84628) from
\url{https://weather.uwyo.edu/upperair/sounding.html} . It can also download
from any other airport by changing the dafault values.}
\usage{
download_airport_sounding(date, hour = 12, airport = "SPIM", region = "samer")
}
\arguments{
\item{date}{Date to download in dd/mm/yyyy.}

\item{hour}{12 or 00. Default is 12}

\item{airport}{Airport code. Default is "SPIM" (Jorge Chavez Airport)}

\item{region}{Region to download. Default is "samer" (South America)}
}
\value{
data.frame with sounding data
}
\description{
Download sounding from Jorge Chavez Airport
Download sounding from Jorge Chavez Airport (SPIM or 84628) from
\url{https://weather.uwyo.edu/upperair/sounding.html} . It can also download
from any other airport by changing the dafault values.
}
\examples{
\dontrun{
# Download sounding from 12z 15/01/2019
jc_sound <- download_airport_sounding("15/01/2019")
}
}
