% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_senamhi_met.R
\name{download_senamhi_met}
\alias{download_senamhi_met}
\title{Download meteorological data from SENAMHI automatic weather stations.}
\usage{
download_senamhi_met(
  aws_codes,
  start_date,
  end_date,
  to_df = FALSE,
  to_csv = FALSE,
  verbose = TRUE,
  csv_path = ""
)
}
\arguments{
\item{aws_codes}{A vector with station codes. See senamhi_aws().}

\item{start_date}{Date to start downloading in dd/mm/yyyy.}

\item{end_date}{Date to end downloading in dd/mm/yyyy.}

\item{to_df}{Returns a data.frame. FALSE by default.}

\item{to_csv}{Export data to csv file. FALSE by default.}

\item{verbose}{Print query summary. TRUE by default.}

\item{csv_path}{Location to export csv.}
}
\value{
list where each element is an air quality station data
}
\description{
This function download the data from automatic weather station (AWS)
described in aws_codes vector. It will return a complete dataset,
this means that missing pollutant data from missing hour is pad out with NA.
}
\examples{
\dontrun{
# Download meteorological data from Campo de Marte and San Borja station
# from 01/02/2022 to 02/02/2022
aws_codes <- c(112181, 112193)
start_date <- "01/02/2022"
end_date <- "02/02/2022"

cm_sb_data <- download_senamhi_met(aws_code, start_date, end_date)
}
}
