% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_senamhi_pol.R
\name{download_senamhi_pol}
\alias{download_senamhi_pol}
\title{Download pollutant data from SENAMHI air quality stations.}
\usage{
download_senamhi_pol(
  aqs_codes,
  pol_codes,
  start_date,
  end_date,
  to_df = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{aqs_codes}{A vector with station codes. See senamhi_aqs().}

\item{pol_codes}{A vector with pollutants codes. See senamhi_params().}

\item{start_date}{Date to start downloading in dd/mm/yyyy.}

\item{end_date}{Date to end downloading in dd/mm/yyyy.}

\item{to_df}{Returns a data.frame. FALSE by default.}

\item{verbose}{Print query summary. TRUE by default.}
}
\value{
list where each element is an air quality station data
}
\description{
This function download many pollutant from pol_codes vector from
the air quality stations described in aqs_codes vector. It will
return a complete dataset, this means that missing pollutant data
from missing hour is pad out with NA.
}
\examples{
\dontrun{
# Download PM10 and PM2.5 from Campo de Marte and San Borja station
# from 01/02/2022 to 02/02/2022
aqs_codes <- c(112194, 112193)
pol_codes <- c("N_PM10", "N_PM25")
start_date <- "01/02/2022"
end_date <- "02/02/2022"

pm_data <- download_senamhi_data(aqs_code, pol_codes, start_date, end_date)
}
}
