% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senamhi_aqs}
\alias{senamhi_aqs}
\title{SENAMHI AQS station latitude and longitude}
\format{
A data frame with 8 observations and 4 variables:
\describe{
\item{aqs}{SENAMHI AQS name.}
\item{code}{SENAMHI AQS code.}
\item{lat}{SENAMHI AQS latitude.}
\item{lon}{SENAMHI AQS longitude.}
}
}
\usage{
senamhi_aqs
}
\description{
SENAMHI air quality stations (AQS) latitudes and longitudes.
Use this to check the AQS \code{aqs_codes} argument in
\code{download_senamhi_pol()}.
#'
}
\examples{
senamhi_aqs
}
\keyword{datasets}
