% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senamhi_aws}
\alias{senamhi_aws}
\title{SENAMHI AWS latitude and longitude}
\format{
A data frame with 11 observations and 4 variables:
\describe{
\item{aqs}{SENAMHI AWS name.}
\item{code}{SENAMHI AWS code.}
\item{lat}{SENAMHI AWS latitude.}
\item{lon}{SENAMHI AWS longitude.}
}
}
\usage{
senamhi_aws
}
\description{
SENAMHI automatic weather station (AWS) latitudes and longitudes.
Use this to check the AQS \code{aws_codes} argument in
\code{download_senamhi_met()}.
#'
}
\examples{
senamhi_aqs
}
\keyword{datasets}
