% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senamhi_params}
\alias{senamhi_params}
\title{SENAMHI pollutant}
\format{
A data frame with 6 observations and 3 variables:
\describe{
\item{code}{SENAMHI pollutant code.}
\item{name}{SENAMHI pollutant name.}
\item{units}{Pollutant unit.}
}
}
\usage{
senamhi_params
}
\description{
SENAMHI air quality stations pollutants.
Use this to check the \code{pol_codes} argument in
\code{download_senamhi_pol()}.
#'
}
\examples{
senamhi_params
}
\keyword{datasets}
